+++ index.html
@@ -38,9 +38,6 @@
             enctype="multipart/form-data"
             action="https://27.javascript.pages.academy/kekstagram-simple"
             autocomplete="off"
-            data-pristine-required-message="Обязательное поле"
-            data-pristine-minlength-message="Минимальная длина 20 символов"
-            data-pristine-maxlength-message="Максимальная длина 140 символов"
           >
 
             <!-- Изначальное состояние поля для загрузки изображения -->
