+++ js/editor/getElements.js
@@ -1,0 +1,13 @@
+export const getSlider = () => document.querySelector('.effect-level__slider');
+
+export const getCurrentFilter = () => document.querySelector('input[name="effect"]:checked').value;
+
+export const getImagePreview = () => document.querySelector('.img-upload__preview > img');
+
+export const getLevelInput = () => document.querySelector('.effect-level__value');
+
+export const getFiltersContainer = () => document.querySelector('.effects__list');
+
+export const getSliderParent = () => document.querySelector('.effect-level');
+
+export const getScaleControl = () => document.querySelector('.scale__control--value');
