+++ js/form/imageModal.js
@@ -1,13 +1,15 @@
-import {addEffectsListeners, removeEffectsListeners} from './effectsEditor.js';
+import {addEffectsListeners, removeEffectsListeners} from '../editor/effectsEditor.js';
 
-export const closeModal = () => {
-  document.querySelector('#upload-select-image').reset();
-  const img = document.querySelector('.img-upload__preview > img');
+export const closeModal = (reset = true) => {
+  if (reset) {
+    document.querySelector('#upload-select-image').reset();
+    const img = document.querySelector('.img-upload__preview > img');
+    img.classList.remove(...img.classList);
+    img.classList.add('effects__preview--none');
+    removeEffectsListeners();
+  }
   document.querySelector('body').classList.remove('modal-open');
   document.querySelector('.img-upload__overlay').classList.add('hidden');
-  img.classList.remove(...img.classList);
-  img.classList.add('effects__preview--none');
-  removeEffectsListeners();
 };
 
 const openModalListener = (e) => {
